pyproject.toml
[project]
name = "mohajer-meal-marketplace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.21.0",
    "anyio==4.9.0",
    "apscheduler==3.11.0",
    "certifi==2025.4.26",
    "greenlet==3.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jalali-core==1.0.0",
    "jdatetime==5.2.0",
    "pillow==11.2.1",
    "python-dotenv==1.1.0",
    "python-telegram-bot==22.0",
    "qrcode==8.2",
    "sniffio==1.3.1",
    "socksio==1.0.0",
    "sqlalchemy==2.0.40",
    "tornado==6.4.2",
    "typing-extensions==4.13.2",
    "tzlocal==5.3.1",
    "alembic>=1.13.1",
    "pytz>=2025.2",
    "python-json-logger>=3.3.0",
]
